Indicate (authenticated) in received header for Postfix.
This change is to help users with Postfix that aren't using the -a option
to not pass mail from authenticated users through SpamAssassin. Postfix,
unlike Sendmail, does not provide the {auth_ssf} macro, so in the case
where mail is from an authenticated user, the modified code can now add
an (authenticated) hint in the dummy Received: header (where Sendmail would
add (authenticated bits=nnn)), and this is scored favourably by SpamAssassin.

http://bugzilla.redhat.com/730308